Update SMD and MG-SMD configs
