+++ packages/devextreme-angular/src/ui/nested/base/popup-toolbar-item-dxi.ts
@@ -1,0 +1,106 @@
+/* tslint:disable:max-line-length */
+
+import { CollectionNestedOption } from 'devextreme-angular/core';
+import {
+    Component,
+} from '@angular/core';
+
+import { ToolbarItemComponent, ToolbarItemLocation } from 'devextreme/common';
+import { ToolbarLocation } from 'devextreme/ui/popup';
+import { LocateInMenuMode, ShowTextMode } from 'devextreme/ui/toolbar';
+
+@Component({
+    template: ''
+})
+export abstract class DxiPopupToolbarItem extends CollectionNestedOption {
+    get cssClass(): string | undefined {
+        return this._getOption('cssClass');
+    }
+    set cssClass(value: string | undefined) {
+        this._setOption('cssClass', value);
+    }
+
+    get disabled(): boolean {
+        return this._getOption('disabled');
+    }
+    set disabled(value: boolean) {
+        this._setOption('disabled', value);
+    }
+
+    get html(): string {
+        return this._getOption('html');
+    }
+    set html(value: string) {
+        this._setOption('html', value);
+    }
+
+    get locateInMenu(): LocateInMenuMode {
+        return this._getOption('locateInMenu');
+    }
+    set locateInMenu(value: LocateInMenuMode) {
+        this._setOption('locateInMenu', value);
+    }
+
+    get location(): ToolbarItemLocation {
+        return this._getOption('location');
+    }
+    set location(value: ToolbarItemLocation) {
+        this._setOption('location', value);
+    }
+
+    get menuItemTemplate(): any {
+        return this._getOption('menuItemTemplate');
+    }
+    set menuItemTemplate(value: any) {
+        this._setOption('menuItemTemplate', value);
+    }
+
+    get options(): any {
+        return this._getOption('options');
+    }
+    set options(value: any) {
+        this._setOption('options', value);
+    }
+
+    get showText(): ShowTextMode {
+        return this._getOption('showText');
+    }
+    set showText(value: ShowTextMode) {
+        this._setOption('showText', value);
+    }
+
+    get template(): any {
+        return this._getOption('template');
+    }
+    set template(value: any) {
+        this._setOption('template', value);
+    }
+
+    get text(): string {
+        return this._getOption('text');
+    }
+    set text(value: string) {
+        this._setOption('text', value);
+    }
+
+    get toolbar(): ToolbarLocation {
+        return this._getOption('toolbar');
+    }
+    set toolbar(value: ToolbarLocation) {
+        this._setOption('toolbar', value);
+    }
+
+    get visible(): boolean {
+        return this._getOption('visible');
+    }
+    set visible(value: boolean) {
+        this._setOption('visible', value);
+    }
+
+    get widget(): ToolbarItemComponent {
+        return this._getOption('widget');
+    }
+    set widget(value: ToolbarItemComponent) {
+        this._setOption('widget', value);
+    }
+}
