+++ tools/ng-smd-cfg.json
@@ -15,5 +15,27 @@
             "kind": "remove",
             "uid": "/stepper:"
         }
-    ]
+    ],
+    "variables": {
+        "collectionItems": [
+            "ui/accordion:dxAccordionItem",
+            "ui/action_sheet:dxActionSheetItem",
+            "ui/box:dxBoxItem",
+            "ui/collection/ui.collection_widget.base:CollectionWidgetItem",
+            "ui/drop_down_button:dxDropDownButtonItem",
+            "ui/gallery:dxGalleryItem",
+            "ui/gantt:dxGanttContextMenuItem",
+            "ui/html_editor:dxHtmlEditorTableContextMenuItem",
+            "ui/list:dxListItem",
+            "ui/menu:dxMenuBaseItem",
+            "ui/multi_view:dxMultiViewItem",
+            "ui/responsive_box:dxResponsiveBoxItem",
+            "ui/scheduler:Appointment",
+            "ui/splitter:dxSplitterItem",
+            "ui/stepper:dxStepperItem",
+            "ui/tab_panel:dxTabPanelItem",
+            "ui/tabs:dxTabsItem",
+            "ui/tile_view:dxTileViewItem"
+        ]
+    }
 }
