+++ tools/smd-cfg.json
@@ -22,6 +22,100 @@
         {
             "kind": "remove",
             "uid": "/scheduler:Toolbar"
+        },
+        {
+            "kind": "remove",
+            "uid": "ui/form:dxFormSimpleItem\\.(editorOptions|editorType)"
+        },
+        {
+            "kind": "add",
+            "uid": "ui/form:dxFormSimpleItem.editor",
+            "types": [
+                {
+                    "kind": "custom",
+                    "name": "WidgetFactory",
+                    "params": {
+                        "factory": "FormItemEditor",
+                        "componentNameProp": "editorType",
+                        "componentConfigProp": "editorOptions"
+                    }
+                }
+            ]
+        },
+        {
+            "kind": "remove",
+            "uid": "ui/toolbar:dxToolbarItem\\.(widget|options)"
+        },
+        {
+            "kind": "add",
+            "uid": "ui/toolbar:dxToolbarItem.widget",
+            "types": [
+                {
+                    "kind": "custom",
+                    "name": "WidgetFactory",
+                    "params": {
+                        "factory": "ToolbarItem",
+                        "componentNameProp": "widget",
+                        "componentConfigProp": "options"
+                    }
+                }
+            ]
+        },
+        {
+            "kind": "add",
+            "uid": "ui/popover:ToolbarItem",
+            "name": "dxPopoverToolbarItem",
+            "parent": "ui/popup:ToolbarItem"
+        },
+        {
+            "kind": "add",
+            "uid": "ui/popover:dxPopoverOptions.toolbarItems",
+            "types": [
+                {
+                    "kind": "array",
+                    "itemTypes": [
+                        {
+                            "kind": "memberRef",
+                            "uid": "ui/popover:ToolbarItem"
+                        }
+                    ]
+                }
+            ]
         }
-    ]
+    ],
+    "variables": {
+        "forwardedEnums": [
+            {
+                "uid": "ui/form:FormItemComponent",
+                "name": "FormItemEditorType"
+            },
+            {
+                "uid": "common:ToolbarItemComponent",
+                "name": "ToolbarItemWidget"
+            }
+        ],
+        "collectionItems": [
+            "ui/accordion:dxAccordionItem",
+            "ui/action_sheet:dxActionSheetItem",
+            "ui/box:dxBoxItem",
+            "ui/card_view:CardHeaderItem",
+            "ui/card_view:ToolbarItem",
+            "ui/collection/ui.collection_widget.base:CollectionWidgetItem",
+            "ui/drop_down_button:dxDropDownButtonItem",
+            "ui/gallery:dxGalleryItem",
+            "ui/gantt:dxGanttContextMenuItem",
+            "ui/html_editor:dxHtmlEditorTableContextMenuItem",
+            "ui/list:dxListItem",
+            "ui/menu:dxMenuBaseItem",
+            "ui/multi_view:dxMultiViewItem",
+            "ui/responsive_box:dxResponsiveBoxItem",
+            "ui/scheduler:Appointment",
+            "ui/scheduler:ToolbarItem",
+            "ui/splitter:dxSplitterItem",
+            "ui/stepper:dxStepperItem",
+            "ui/tab_panel:dxTabPanelItem",
+            "ui/tabs:dxTabsItem",
+            "ui/tile_view:dxTileViewItem"
+        ]
+    }
 }
